+++ src/main/java/com/phantomwing/choppersdelight/ui/ModCreativeModeTab.java
@@ -2,6 +2,7 @@ package com.phantomwing.choppersdelight.ui;
 
 import com.phantomwing.choppersdelight.ChoppersDelight;
 import com.phantomwing.choppersdelight.Compatibility;
+import com.phantomwing.choppersdelight.EveryCompatSetup;
 import com.phantomwing.choppersdelight.block.ModBlocks;
 import com.phantomwing.choppersdelight.component.ModDataComponents;
 import com.phantomwing.choppersdelight.item.ModItems;
@@ -69,6 +70,7 @@ public class ModCreativeModeTab {
         displayModItems(parameters, output);
         displayBiomesOPlentyItems(parameters, output);
         displayBiomesWeveGoneItems(parameters, output);
+        displayEveryCompatItems(parameters, output);
 
         // Add some preconfigured designs.
         output.accept(getDecoratedCuttingBoard(ModBlocks.CHERRY_CUTTING_BOARD, Items.WHITE_BANNER, BannerPatterns.FLOWER, DyeColor.PINK));
@@ -110,6 +112,15 @@ public class ModCreativeModeTab {
         ModItems.BIOMES_WEVE_GONE_CREATIVE_TAB_ITEMS.forEach((item) -> {
             output.accept(item.get());
         });
+    }
+
+    public static void displayEveryCompatItems(CreativeModeTab.ItemDisplayParameters parameters, CreativeModeTab.Output output) {
+        if (!Compatibility.IsEveryCompatLoaded()) {
+            return;
+        }
+
+        // Add items to this tab.
+        EveryCompatSetup.forEachCuttingBoardItem(output::accept);
     }
 
     public static void register(IEventBus eventBus) {